debian: work-around multi-arch usability
We need a work-around to enable mesa cross-compiling on CI for it's
taking too much time to do full native builds. From Debian bug 769380 we
know this is actually a unresolved problem about interpretor multiarch
implementation, but I just need this badly so the best work-arounds so
far is to mark python-mako as multi-arch foreign and remove
python-markupsafe completely from dependecy chain.

[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=769380
[2]: https://wiki.debian.org/HelmutGrohne/MultiarchSpecChanges#Interpreter_issue

Closes: https://github.com/laarid/package_debian-mesa/issues/2